requirements.txt
streamlit>=1.28.0
openai>=1.0.0
anthropic>=0.7.0
tiktoken>=0.5.0
jsonschema>=4.19.0
jsonref>=1.1.0
nltk>=3.8.0
spacy>=3.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.1.0
# spaCy language model (mandatory for enhanced mode)
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl